KL-176/refactor: apply CodeRabbit suggestions
