+++ src/constants/defaultWallpaper.js
@@ -1,0 +1,4 @@
+const defaultWallpaper =
+  'https://image.fmkorea.com/files/attach/new3/20231020/3655109/5782408085/6302086124/a35751389c84e91bfa145bce6f81e5ba.jpg'
+
+export default defaultWallpaper
