+++ src/pages/LogoutPage/LogoutPage.jsx
@@ -13,9 +13,9 @@ function LogoutPage() {
   useEffect(() => {
     const logout = async () => {
       if (isLogin) {
-        setLogout()
         try {
           await mutateAsync()
+          setLogout()
         } catch (error) {
           alert('로그아웃에 실패했습니다. 다시 시도해주세요.')
         }
